backend/requirements.txt
flask =3.0.0
flask-cors =3.0.10
PyPDF2 =3.0.1
python-docx = 1.1.0